+++ game/server_core.cpp
@@ -35,7 +35,7 @@ void ServerCore::run()
             }
         }
         receive_data();
-        process_input();
+        // process_input(); Moved to receive_data - called per packet
         update_game_state();
         send_updates();
     }
@@ -75,6 +75,7 @@ void ServerCore::receive_data()
             {
                 InputPacket packet;
                 InputPacket::deserialize(buf, packet);
+                process_input(packet);
 
                 // Print for testing
                 printf("\nEvents: ");
@@ -87,15 +88,49 @@ void ServerCore::receive_data()
     }
 }
 
-void ServerCore::process_input() {}
+void ServerCore::process_input(InputPacket packet) {
+    // For now operate on first player by default. TODO: Identify by socket num? Client ID?
+    glm::mat4 world = serverState.players[0].world;
+
+    float SCALE = 0.05f; // TODO: Define this somewhere else. Maybe in a constants folder?
+
+    // Process input events
+    for (int event : packet.events) {
+        glm::vec3 dir;
+        switch (event) {
+        case MOVE_FORWARD:
+            dir = glm::vec3(0.0f, 0.0f, -1.0f);
+            break;
+        case MOVE_BACKWARD:
+            dir = glm::vec3(0.0f, 0.0f, 1.0f);
+			break;
+        case MOVE_LEFT:
+			dir = glm::vec3(-1.0f, 0.0f, 0.0f);
+			break;
+        case MOVE_RIGHT:
+            dir = glm::vec3(1.0f, 0.0f, 0.0f);
+            break;
+        }
+
+        // Rotate dir by camera angle
+        dir = glm::normalize(glm::rotateY(dir, packet.cam_angle));
+
+        world = glm::translate(world, dir * SCALE);
+
+    }
+
+    serverState.players[0].world = world;
+}
 
 void ServerCore::update_game_state() {
+
+    // Update parts of the game that don't depend on player input.
+
+    // Enemy AI etc
     while (serverState.students.size() < 5) {
         StudentState s_state;
-        s_state.x = serverState.students.size();
-        s_state.y = serverState.students.size();
-        s_state.z = serverState.students.size();
-        s_state.orientation = serverState.students.size();
+
+        s_state.world = glm::mat4(1.0f); // TEMP
 
         serverState.students.push_back(s_state);
     }
@@ -133,10 +168,8 @@ void ServerCore::accept_new_clients()
     clients_data.push_back(client);
 
     PlayerState p_state;
-    p_state.x = 0.0f;
-    p_state.y = 0.0f;
-    p_state.z = 0.0f;
-    p_state.orientation = 0.0f;
+    p_state.world = glm::mat4(1.0f);
+
     p_state.score = 0;
 
     serverState.players.push_back(p_state);
