+++ include/.vs/VSWorkspaceState.json
@@ -1,0 +1,6 @@
+{
+  "ExpandedNodes": [
+    ""
+  ],
+  "PreviewInSolutionExplorer": false
+}
