+++ include/core.h
@@ -4,7 +4,6 @@
  * Graphics includes - used in most files
  * + Any common definitions / global variables we might need
  */
-
 #include <glad/glad.h>
 // Include GLAD before other header files that require openGL (like GLFW!)
 #include <GLFW/glfw3.h>
