+++ include/server_core.h
@@ -13,6 +13,9 @@
 #include "game_state_packet.h"
 #include "game_state.h"
 
+#define GLM_ENABLE_EXPERIMENTAL
+#include <glm/gtx/rotate_vector.hpp>
+
 #define NUM_CLIENTS 1
 
 struct ClientData {
@@ -34,7 +37,7 @@ class ServerCore {
         bool isRunning() const;         // Check if the server is running
 
         void receive_data();            // Receive data from clients
-        void process_input();           // Process inputs
+        void process_input(InputPacket packet);           // Process inputs
         void update_game_state();       // Update the game state
         void send_updates();            // Send updates to clients
         void accept_new_clients();
@@ -43,8 +46,6 @@ class ServerCore {
         Server server;
         std::vector<ClientData> clients_data;
         GameState serverState;
-        // std::vector<ClientData> data;   // all client data passed in. later add other data like changes in npc or environment?
-        std::vector<std::string> data;  // string for now
 };
 
 #endif
